Fix: change result value type

+++ src/CalculatorTests.py
@@ -32,7 +32,7 @@ class MyTestCase(unittest.TestCase):
     def test_divide_method_calculator(self):
         test_data = CsvReader("src/Unit Test Division.csv").data
         for row in test_data:
-            self.assertEqual(self.calculator.divide(row['Value 1'], row['Value 2']), int(row['Result']))
+            self.assertEqual(self.calculator.divide(row['Value 1'], row['Value 2']), float(row['Result']))
             self.assertEqual(self.calculator.result, float(row['Result']))
